Notifications: use std::string instead of strchr because of const char*
The variable `title` is defined as `const char*`, which means, that
`strchr()` returns a `const char*` as well according to
https://www.cplusplus.com/reference/cstring/strchr/

But in the same line the return value is assigned to a non-const
`char*`, which shouldn't be allowed (error with `-pedantic`).

To prevent manual memory management just use `std::string` to copy the
provided title and replace newlines in the copied string.